+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 [Ludusavi v0.7.0 or newer](https://github.com/mtkennerly/ludusavi/releases) is now required.
 
+* Added:
+  * "Browse" button for Ludusavi executable file.
+  * "Browse" and "open" buttons for backup directory.
 * Fixed:
   * Some error conditions reported by Ludusavi were not handled properly.
 * Changed:
@@ -12,6 +15,8 @@
     If the update doesn't work (such as when your Internet is down), the plugin
     will still operate as normal.
   * File size units are now adjusted based on the size, rather than always using MiB.
+  * The default backup directory now writes out the user folder in full rather
+    than using the `~` placeholder (although that is still supported).
 
 ## v0.1.0 (2020-07-29)